+++ HarpGuru/HarpFace/HarpFaceFragment/styles/index.ts
@@ -6,7 +6,7 @@ import type { HarpFaceFragmentProps, HarpFaceFragmentStyles } from '../types'
 import { getFragmentFacts } from '../getFragmentFacts'
 import { columnWidth, rowHeight } from '../../styles'
 import { getHarpFaceFacts } from '../../helpers'
-import type { SetActiveHarpStrata } from '../../../types'
+import type { SetActiveHarpStrata } from '../../../helpers'
 
 type Props = HarpFaceFragmentProps & {
   readonly activeHarpStrata: HarpStrata
