+++ HarpGuru/helpers/index.ts
@@ -1,3 +1,4 @@
 export { usePrevious } from './usePrevious'
 export { getPropsForHarpStrata } from './getPropsForHarpStrata'
 export { setGlobalReactNState } from './setGlobalReactNState'
+export type { SetActiveHarpStrata } from './setGlobalReactNState'
